Refactor pdf generation code outside of index.js
To separate pdf generation from the Express server.
